Move algorithm info list and structs to a dedicated file
